pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-chatbot"
version = "0.1.0"
description = "FastAPI RAG app with a simple frontend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "langchain>=0.3.0",
  "langchain-google-genai>=2.0.0",
  "langchain-community>=0.3.0",
  "langchain-text-splitters>=0.3.0",
  "langchain-core>=0.3.0",
  "docx2txt>=0.8",
  "pypdf",
  "python-multipart",
  "streamlit",
  "python-dotenv",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "rag*", "src*"]
